delete one audio

+++ app/user/word.py
@@ -49,34 +49,36 @@ class RecordView(Resource):
 
 class UpdateRecordView(Resource):
     def get(self, word_id):
-        return {'hello': word_id}
+        return {'word_id': word_id}
 
     def post(self, word_id):
-
         args = parser.parse_args()
         basepath = os.path.abspath(os.getcwd())
         userid = User.decode_auth_token(args['token'])
         userwordrel = Userwordrel.query.filter_by(user_id=userid, word_id=word_id).first()
-        try:
-            stream = args.get('record')
-            upload_path = os.path.join(basepath, "records", secure_filename(stream.filename))
-            stream.save(upload_path)
-            userwordrel.update_at = datetime.now()
-            userwordrel.snd_abs = args['snd_abs']
-            userwordrel.verified = True
-        except FileNotFoundError as e:
-            current_app.logger.error(e)
-            return {"status": 400, "message": "未上传录音文件"}, 400
-        try:
-            db.session.commit()  # SQLAlchemy用
-        except Exception as e:
-            # 数据库出错回滚
-            db.session.rollback()
-            current_app.logger.error(e)
-            os.remove(upload_path)
-            return {"status": 404, "message": "数据库查询异常"}, 404
+        if userwordrel is not None:
+            try:
+                stream = args.get('record')
+                upload_path = os.path.join(basepath, "records", secure_filename(stream.filename))
+                stream.save(upload_path)
+                userwordrel.update_at = datetime.now()
+                userwordrel.snd_abs = args['snd_abs']
+                userwordrel.verified = True
+            except FileNotFoundError as e:
+                current_app.logger.error(e)
+                return {"status": 400, "message": "未上传录音文件"}, 400
+            try:
+                db.session.commit()  # SQLAlchemy用
+            except Exception as e:
+                # 数据库出错回滚
+                db.session.rollback()
+                current_app.logger.error(e)
+                os.remove(upload_path)
+                return {"status": 404, "message": "数据库查询异常"}, 404
 
-        return {"status": 200, "message": "upload success"}, 200
+            return {"status": 200, "message": "upload success"}, 200
+        else:
+            return {'status': 403, 'message': 'no this audio'}, 403
 
 
 user_api.add_resource(RecordView, '/<word_id>/record')
